GRIFFIN: renaming packages/classes to have names conistent with Lexer/Parser approach for language compilers.
