+++ core/src/main/java/com/questdb/griffin/lexer/ExprAstBuilder.java
@@ -21,10 +21,47 @@
  *
  ******************************************************************************/
 
-package com.questdb.griffin.parser;
+package com.questdb.griffin.lexer;
 
 import com.questdb.griffin.common.ExprNode;
 
-public interface ExprListener {
-    void onNode(ExprNode node);
+import java.util.ArrayDeque;
+import java.util.Deque;
+
+public final class ExprAstBuilder implements ExprListener {
+
+    private final Deque<ExprNode> stack = new ArrayDeque<>();
+
+    @Override
+    public void onNode(ExprNode node) {
+        switch (node.paramCount) {
+            case 0:
+                break;
+            case 1:
+                node.rhs = stack.poll();
+                break;
+            case 2:
+                node.rhs = stack.poll();
+                node.lhs = stack.poll();
+                break;
+            default:
+                for (int i = 0; i < node.paramCount; i++) {
+                    node.args.add(stack.poll());
+                }
+                break;
+        }
+        stack.push(node);
+    }
+
+    public ExprNode poll() {
+        return stack.poll();
+    }
+
+    public void reset() {
+        stack.clear();
+    }
+
+    public int size() {
+        return stack.size();
+    }
 }
