+++ core/src/main/java/com/questdb/griffin/lexer/IntervalCompiler.java
@@ -21,7 +21,7 @@
  *
  ******************************************************************************/
 
-package com.questdb.griffin.parser;
+package com.questdb.griffin.lexer;
 
 import com.questdb.common.NumericException;
 import com.questdb.std.LongList;
